# gh-121562: optimized hex_from_char