--- a/Misc/NEWS.d/next/Core and Builtins/2024-07-10-02-02-32.gh-issue-121562.8beIMi.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2024-07-10-02-02-32.gh-issue-121562.8beIMi.rst
@@ -1,0 +1,2 @@
+Optimized performance of hex_from_char by replacing switch-case with a
+lookup table