# Optimized hex_from_char performance

issue https://github.com/python/cpython/issues/121562

Timing with old hex_from_char:

```
$ ./python -m pyperf timeit --duplicate 100 "float.fromhex('0x123456.ffffp10')"
.....................
Mean +- std dev: 86.0 ns +- 0.9 ns
```

With new hex_from_char:

```
$ ./python -m pyperf timeit --duplicate 100 "float.fromhex('0x123456.ffffp10')"
.....................
Mean +- std dev: 81.8 ns +- 1.0 ns
```

If we measure `hex_from_char` directly, probably we will get a much better % improvement because there is additional calls that `float.fromhex` makes apart from this one.

<!-- gh-issue-number: gh-121562 -->
* Issue: gh-121562
<!-- /gh-issue-number -->